eliminate unnecessary NxtClusterDB

+++ db/cluster.go
@@ -26,21 +26,18 @@ const maxClusters = 100 // A swag
 // operational data relevant to just that cluster.
 
 var mongoClient *mongo.Client
-var globalClusterDB *mongo.Database
 var ClusterDBs = make(map[string]*mongo.Database, maxClusters)
 
-var clusterGwCltn *mongo.Collection
-var usersCltn = make(map[string]*mongo.Collection, maxClusters)
+var tenantsCltn = make(map[string]*mongo.Collection, maxClusters)
 var bundleCltn = make(map[string]*mongo.Collection, maxClusters)
+var gatewaysCltn = make(map[string]*mongo.Collection, maxClusters)
 
 func ClusterDBInit(dbClient *mongo.Client) {
 	mongoClient = dbClient
-	globalClusterDB = dbClient.Database("NxtClusterDB")
-	clusterGwCltn = globalClusterDB.Collection("NxtGateways")
 }
 
 func ClusterGetDBName(cl string) string {
-	return ("Nxt-" + cl + "-DB")
+	return ("Cluster-" + cl + "-DB")
 }
 
 func ClusterGetApodSetName(tenant string, pod int) string {
@@ -55,11 +52,17 @@ func ClusterGetCollection(cluster string, cltn string) *mongo.Collection {
 	}
 	switch cltn {
 	case "NxtTenants":
-		_, cok := usersCltn[cluster]
+		_, cok := tenantsCltn[cluster]
 		if cok == false {
-			usersCltn[cluster] = ClusterDBs[cluster].Collection("NxtTenants")
+			tenantsCltn[cluster] = ClusterDBs[cluster].Collection("NxtTenants")
 		}
-		return usersCltn[cluster]
+		return tenantsCltn[cluster]
+	case "NxtGateways":
+		_, cok := gatewaysCltn[cluster]
+		if cok == false {
+			gatewaysCltn[cluster] = ClusterDBs[cluster].Collection("NxtGateways")
+		}
+		return gatewaysCltn[cluster]
 	case "NxtConnectors":
 		_, cok := bundleCltn[cluster]
 		if cok == false {
@@ -81,12 +84,13 @@ func ClusterAddDB(cluster string) {
 }
 
 func ClusterAddCollections(cluster string, cldb *mongo.Database) {
-	usersCltn[cluster] = cldb.Collection("NxtTenants")
+	tenantsCltn[cluster] = cldb.Collection("NxtTenants")
 	bundleCltn[cluster] = cldb.Collection("NxtConnectors")
+	gatewaysCltn[cluster] = cldb.Collection("NxtGateways")
 }
 
 func ClusterDelDB(cluster string) {
-	delete(usersCltn, cluster)
+	delete(tenantsCltn, cluster)
 	delete(bundleCltn, cluster)
 	ClusterDBs[cluster].Drop(context.TODO())
 	delete(ClusterDBs, cluster)
@@ -102,7 +106,6 @@ func ClusterDBDrop() error {
 		cldb := mongoClient.Database(ClusterGetDBName(Cluster))
 		cldb.Drop(context.TODO())
 	}
-	globalClusterDB.Drop(context.TODO())
 
 	return nil
 }
@@ -133,6 +136,13 @@ func DBAddClusterGateway(data *Gateway, remotes []string) error {
 		ReturnDocument: &after,
 		Upsert:         &upsert,
 	}
+	clusterGwCltn := ClusterGetCollection(Cluster, "NxtGateways")
+	if clusterGwCltn == nil {
+		msg := fmt.Sprintf("Could not find gw collection for cluster %s",
+			Cluster)
+		glog.Error(msg)
+		return errors.New(msg)
+	}
 	err := clusterGwCltn.FindOneAndUpdate(
 		context.TODO(),
 		bson.M{"_id": data.Name},
@@ -155,6 +165,14 @@ func DBAddClusterGateway(data *Gateway, remotes []string) error {
 // Find gateway/cluster doc given the gateway name
 func DBFindGatewayCluster(gwname string) (error, *ClusterGateway) {
 	var gateway ClusterGateway
+	Cluster := DBGetClusterName(gwname)
+	clusterGwCltn := ClusterGetCollection(Cluster, "NxtGateways")
+	if clusterGwCltn == nil {
+		msg := fmt.Sprintf("Could not find gw collection for cluster %s",
+			Cluster)
+		glog.Error(msg)
+		return errors.New(msg), nil
+	}
 	err := clusterGwCltn.FindOne(
 		context.TODO(),
 		bson.M{"_id": gwname},
@@ -175,6 +193,14 @@ func DBDelClusterGateway(gwname string) error {
 	}
 	if clgw == nil {
 		return errors.New("Cluster gateway not found - cannot delete")
+	}
+	Cluster := DBGetClusterName(gwname)
+	clusterGwCltn := ClusterGetCollection(Cluster, "NxtGateways")
+	if clusterGwCltn == nil {
+		msg := fmt.Sprintf("Could not find gw collection for cluster %s",
+			Cluster)
+		glog.Error(msg)
+		return errors.New(msg)
 	}
 	_, err := clusterGwCltn.DeleteOne(
 		context.TODO(),
@@ -221,6 +247,13 @@ func DBAddDelClusterGatewayRemote(thisGw string, remoteGw string, add bool) erro
 	opt := options.FindOneAndUpdateOptions{
 		ReturnDocument: &after,
 		Upsert:         &upsert,
+	}
+	clusterGwCltn := ClusterGetCollection(gw.Cluster, "NxtGateways")
+	if clusterGwCltn == nil {
+		msg := fmt.Sprintf("Could not find gw collection for cluster %s",
+			gw.Cluster)
+		glog.Error(msg)
+		return errors.New(msg)
 	}
 	err := clusterGwCltn.FindOneAndUpdate(
 		context.TODO(),
